Split perf_test_utils
